modularize and nest "calc_from_args_utils" and "epp_utils"
