+++ scripts/ont_sync_to_db.py
@@ -18,7 +18,7 @@ from genologics.entities import Artifact, Process
 from genologics.lims import Lims
 from ont_send_reloading_info_to_db import get_ONT_db
 
-from epp_utils import udf_tools
+from scilifelab_epps.epp_utils import udf_tools
 from scilifelab_epps.epp import upload_file
 
 DESC = """Script for finishing the step to start ONT sequencing in LIMS.
