+++ scripts/parse_ba_results.py
@@ -11,7 +11,7 @@ from genologics.config import BASEURI, PASSWORD, USERNAME
 from genologics.entities import Process
 from genologics.lims import Lims
 
-from epp_utils import udf_tools
+from scilifelab_epps.epp_utils import udf_tools
 from scilifelab_epps.epp import get_well_number
 
 DESC = """This script parses the Agilent BioAnalyzer XML report.
