+++ scripts/qc_amount_calculation.py
@@ -16,7 +16,7 @@ from genologics.config import BASEURI, PASSWORD, USERNAME
 from genologics.entities import Process
 from genologics.lims import Lims
 
-from epp_utils import formula, udf_tools
+from scilifelab_epps.epp_utils import formula, udf_tools
 from scilifelab_epps.epp import EppLogger
