+++ scripts/zika_methods.py
@@ -11,7 +11,7 @@ import sys
 import numpy as np
 import pandas as pd
 
-from epp_utils.udf_tools import is_filled
+from scilifelab_epps.epp_utils.udf_tools import is_filled
 from scripts import zika_utils
